implementacion de size_t a uint32_t, se cambio page y pageManager

+++ RENIEC_api/src/page.h
@@ -12,17 +12,17 @@
 #include <cstdint>
 #include <cmath>
 
-const size_t MAX_KEYS = 120;
-const size_t PAGE_SIZE = 4096; //tam de pagina 4kb
-const size_t MIN_KEYS = static_cast<size_t>(std::ceil((2.0 * MAX_KEYS) / 3.0));
+const uint32_t MAX_KEYS = 120;
+const uint32_t PAGE_SIZE = 4096; //tam de pagina 4kb
+const uint32_t MIN_KEYS = static_cast<uint32_t>(std::ceil((2.0 * MAX_KEYS) / 3.0));
 
 
 struct Page {
     bool is_leaf;
-    size_t num_keys;
+    uint32_t num_keys;
     IndexEntry entries[MAX_KEYS];
-    size_t children[MAX_KEYS + 1]; // Identificadores de páginas hijas
-    size_t parent;                 // Identificador de la página padre
+    uint32_t children[MAX_KEYS + 1]; // Identificadores de páginas hijas
+    uint32_t parent;                 // Identificador de la página padre
 
     // Constructor por defecto
     Page() : is_leaf(true), num_keys(0), parent(0) {
